third_party/blink/web_tests/platform/generic/svg/dom/SVGLocatable-getCTM-svg-root-expected.txt

This test checks getCTM() on SVG root elements with borders, paddings, and margins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getCTMById('svgRoot1') is "[1 0 0 1 0 0]"
PASS getCTMById('svgRoot2') is "[0.5 0 0 0.5 0 0]"
PASS getCTMById('svgRoot3') is "[2 0 0 2 0 0]"
PASS getCTMById('svgRoot4') is "[1 0 0 1 0 0]"
PASS getCTMById('svgRoot5') is "[1 0 0 1 0 0]"
PASS getCTMById('svgRoot6') is "[0.5 0 0 0.5 0 0]"
PASS successfullyParsed is true

TEST COMPLETE
